backend/pyproject.toml
[project]
name = "paradise-voice-agent"
version = "0.1.0"
description = "RAG-enabled travel voice agent with LiveKit"
requires-python = ">=3.10"
dependencies = [
    "livekit-agents[openai,silero,turn-detector]~=1.2",
    "livekit>=0.11.0",
    "livekit-plugins-langchain>=0.1.0",
    "python-dotenv>=1.0.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-pinecone>=0.1.0",
    "pinecone-client>=3.0.0",
    "pypdf>=4.0.0",
    "requests>=2.31.0",
    "googlemaps>=4.10.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

